create unittests for file input retriever
