+++ genai-perf/tests/test_input_retriever_factory.py
@@ -12,21 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from collections import namedtuple
-from pathlib import Path
-from unittest.mock import mock_open, patch
+from unittest.mock import patch
 
 import pytest
 import responses
 from genai_perf.exceptions import GenAIPerfException
 from genai_perf.inputs import input_constants as ic
-from genai_perf.inputs.file_input_retriever import FileInputRetriever
-from genai_perf.inputs.input_constants import ModelSelectionStrategy, OutputFormat
+from genai_perf.inputs.input_constants import OutputFormat
 from genai_perf.inputs.input_retriever_factory import InputRetrieverFactory
 from genai_perf.inputs.inputs_config import InputsConfig
 from genai_perf.inputs.synthetic_image_generator import ImageFormat
 from genai_perf.tokenizer import DEFAULT_TOKENIZER, get_tokenizer
-from PIL import Image
 
 mocked_openorca_data = {
     "features": [
@@ -197,89 +193,3 @@ class TestInputRetrieverFactory:
 
         assert dataset_json is not None
         assert len(dataset_json["rows"]) == TEST_LENGTH
-
-    def test_get_input_file_without_file_existing(self):
-        file_retriever = FileInputRetriever(
-            InputsConfig(input_filename=Path("prompt.txt"))
-        )
-        with pytest.raises(FileNotFoundError):
-            file_retriever._get_input_dataset_from_file()
-
-    @patch("pathlib.Path.exists", return_value=True)
-    @patch(
-        "builtins.open",
-        new_callable=mock_open,
-        read_data='{"text_input": "single prompt"}\n',
-    )
-    def test_get_input_file_with_single_prompt(self, mock_file, mock_exists):
-        expected_prompts = ["single prompt"]
-        file_retriever = FileInputRetriever(
-            InputsConfig(
-                model_name=["test_model_A"],
-                model_selection_strategy=ModelSelectionStrategy.ROUND_ROBIN,
-                input_filename=Path("prompt.txt"),
-            )
-        )
-        dataset = file_retriever._get_input_dataset_from_file()
-
-        assert dataset is not None
-        assert len(dataset["rows"]) == len(expected_prompts)
-        for i, prompt in enumerate(expected_prompts):
-            assert dataset["rows"][i]["row"]["text_input"] == prompt
-
-    @patch("pathlib.Path.exists", return_value=True)
-    @patch(
-        "builtins.open",
-        new_callable=mock_open,
-        read_data='{"text_input": "prompt1"}\n{"text_input": "prompt2"}\n{"text_input": "prompt3"}\n',
-    )
-    def test_get_input_file_with_multiple_prompts(self, mock_file, mock_exists):
-        expected_prompts = ["prompt1", "prompt2", "prompt3"]
-        file_retriever = FileInputRetriever(
-            InputsConfig(
-                model_name=["test_model_A"],
-                model_selection_strategy=ModelSelectionStrategy.ROUND_ROBIN,
-                input_filename=Path("prompt.txt"),
-            )
-        )
-        dataset = file_retriever._get_input_dataset_from_file()
-
-        assert dataset is not None
-        assert len(dataset["rows"]) == len(expected_prompts)
-        for i, prompt in enumerate(expected_prompts):
-            assert dataset["rows"][i]["row"]["text_input"] == prompt
-
-    @patch("pathlib.Path.exists", return_value=True)
-    @patch("PIL.Image.open", return_value=Image.new("RGB", (10, 10)))
-    @patch(
-        "builtins.open",
-        new_callable=mock_open,
-        read_data=(
-            '{"text_input": "prompt1", "image": "image1.png"}\n'
-            '{"text_input": "prompt2", "image": "image2.png"}\n'
-            '{"text_input": "prompt3", "image": "image3.png"}\n'
-        ),
-    )
-    def test_get_input_file_with_multi_modal_data(
-        self, mock_exists, mock_image, mock_file
-    ):
-        file_retriever = FileInputRetriever(
-            InputsConfig(
-                model_name=["test_model_A"],
-                model_selection_strategy=ModelSelectionStrategy.ROUND_ROBIN,
-                input_filename=Path("prompt.txt"),
-            )
-        )
-        Data = namedtuple("Data", ["text_input", "image"])
-        expected_data = [
-            Data(text_input="prompt1", image="image1.png"),
-            Data(text_input="prompt2", image="image2.png"),
-            Data(text_input="prompt3", image="image3.png"),
-        ]
-        dataset = file_retriever._get_input_dataset_from_file()
-
-        assert dataset is not None
-        assert len(dataset["rows"]) == len(expected_data)
-        for i, data in enumerate(expected_data):
-            assert dataset["rows"][i]["row"]["text_input"] == data.text_input
-            assert dataset["rows"][i]["row"]["image"] == data.image
